Use mock-docs in techdocs-container for development.
Enabling the plugin causes the live-reload feature of mkdocs to crash.
However,  only crashes in the example-docs directory in this project.
Works well in standalone directories and example docs. Needs investigation.
Could be related to the temporary files created by the mkdocs-techdocs-core plugin
and them being removed after creation, but live-reload wants to track them.